function_api/requirements.txt
azure-functions
requests
pydantic
azure-storage-blob
datetime
